Added basic Dockerfile
Added a basic Dockerfile running an older version of Metabase that was
released prior to the update that supported the latests Google
authentication libraries.

If the test works as expected. We should not be able to configure Google
Sign-In against this version.

If we can then upgrade to the latest version of Metabase and everything
works, then I would feel confident we are not going to hit issues when
Google removes support for the JavaScript platform library.

https://developers.googleblog.com/2021/08/gsi-jsweb-deprecation.html